+++ dummy-project/src/main/kotlin/com/labijie/orm/dummy/pojo/dsl/UserDSL.kt
@@ -242,8 +242,8 @@ public object UserDSL {
     offsetKey?.let {
       when(order) {
         SortOrder.DESC, SortOrder.DESC_NULLS_FIRST, SortOrder.DESC_NULLS_LAST->
-        query.andWhere { id less it }
-        else-> query.andWhere { id greater it }
+        query.andWhere { id less it.toLong() }
+        else-> query.andWhere { id greater it.toLong() }
       }
     }
     `where`?.invoke(query)
